pyproject.toml
[project]
name = "fin-track"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "mypy>=1.19.1",
    "pandas>=3.0.0",
    "pandera>=0.28.1",
    "pyaml>=25.7.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "rapidfuzz>=3.14.3",
    "sentence-transformers>=5.2.0",
    "torch>=2.10.0",
]
